fix: Restructure type hierarchy so that types don't conflict internally.
The old `Result<TValue, TError> = ResultValue<TValue> | ResultError<TError>`
lead to the expansion `ResultBase<TValue, Error> | ResultBase<unknown, TError>`
which lead to a conflict on the `unwrapOrThrow` method and others, where
the return type was suddenly `TValue | unknown` and thus always unknown.

This commit restructures the type hierarchy so that this conflict is
avoided.